+++ src/main/java/bsuir/vintsarevich/buisness/product/service/impl/ProductService.java
@@ -102,7 +102,7 @@ public class ProductService implements IProductService {
     }
 
     @Override
-    public void addProduct(String type, String nameRu, String nameEn, Integer weight, Double cost, String status,
+    public void addProduct(String type, String nameRu, String nameEn, Double weight, Double cost, String status,
                            String descriptionRu, String descriptionEn, Part image, String webPath) throws ServiceException, ServiceLogicException {
         LOGGER.log(Level.DEBUG, "ProductService: addProduct start");
         Product product = new Product();
